+++ src/components/PlayStopButton.tsx
@@ -3,9 +3,9 @@ import { CaretRightFilled, PauseOutlined } from '@ant-design/icons';
 import { observer } from 'mobx-react';
 import { createUseStyles } from 'react-jss';
 
-import CircleButton from '../CircleButton/CircleButton';
-import rootStore from '../../modules/RootStore';
-import TaskModel from '../../modules/tasks/models/TaskModel';
+import CircleButton from './CircleButton';
+import rootStore from '../modules/RootStore';
+import TaskModel from '../modules/tasks/models/TaskModel';
 
 const { tasksStore } = rootStore;
