+++ src/components/TimeRangeModal/TimeRangeModal.tsx
@@ -10,7 +10,7 @@ import rootStore from '../../modules/RootStore';
 import TaskTimeItemModel from '../../modules/tasks/models/TaskTimeItemModel';
 import { ITimeRangeModel } from '../../modules/tasks/models/TaskModel';
 import { Undefined } from '../../types/CommonTypes';
-import TimeRangeDuration from './components/TimeRangeDuration';
+import TimeRangeDuration from './TimeRangeDuration';
 import IModalProps from '../../types/IModalProps';
 
 const { tasksStore } = rootStore;
