+++ src/screens/dashboard/Dashboard.tsx
@@ -3,7 +3,7 @@ import { createUseStyles } from 'react-jss';
 import { Layout, Space } from 'antd';
 import { observer } from 'mobx-react';
 
-import SelectDate from '../../components/SelectDate/SelectDate';
+import SelectDate from '../../components/SelectDate';
 import rootStore from '../../modules/RootStore';
 import { getTasksWithTotalTimeForDay } from '../../helpers/TaskHelper';
 import HoursWithDuration from './components/HoursWithDuration';
